+++ tests/files/table-comments/input.lua
@@ -1,5 +1,5 @@
-local x =  {
-    foo, -- comment
-    bar, -- test
-    baz
+local x = {
+	"foo", -- comment
+	"bar", -- test
+	"baz" -- test
 }